requirements.txt
KindleComicConverter >= 5.5.2
ruamel.std.zipfile >= 0.1.0
psutil ~= 5.7.3